Add Workspace Modal as a separate component
Refactored code to move Add Workspace Modal to its own components.
